Add riscv32imc and riscv64gc to the known sha targets
Add the representative riscv32imc and riscv64gc targets to the list of
known targets to populate the known shas list:

- riscv32imc-unknown-none-elf
- riscv64gc-unknown-none-elf

Run `//util:fetch_shas` to update the known shas list. This adds the
rust-std archive hashes for these targets for the supported stable,
beta, and nightly releases to the `FILE_KEY_TO_SHA` dict so these hashes
will be available to `load_arbitrary_tool()` to allow Bazel to cache the
downloaded archives.

Amended commit to bump CI.